code/autograder/exercises/ex06/dictionary_test.py
"""A mock unit test file for dictionaries."""

from exercises.ex06.dictionary import invert, favorite_color, count, alphabetizer, update_attendance


def test_invert_1() -> None:
    """Mock unit tests for dictionaries."""
    assert invert({"apple": "cat"}) == {"cat": "apple"}


def test_invert_2() -> None:
    """Mock unit tests for dictionaries."""
    assert invert({"apple": "cat"}) == {"cat": "apple"}


def test_invert_3() -> None:
    """Mock unit tests for dictionaries."""
    assert invert({"apple": "cat"}) == {"cat": "apple"}


def test_favorite_color_1() -> None:
    """Mock unit tests for dictionaries."""
    assert favorite_color({"Bob": "blue"}) == "blue"


def test_favorite_color_2() -> None:
    """Mock unit tests for dictionaries."""
    assert favorite_color({"Bob": "blue"}) == "blue"


def test_favorite_color_3() -> None:
    """Mock unit tests for dictionaries."""
    assert favorite_color({"Bob": "blue"}) == "blue"


def test_count_1() -> None:
    """Mock unit tests for dictionaries."""
    assert count(["apple"]) == {"apple": 1}


def test_count_2() -> None:
    """Mock unit tests for dictionaries."""
    assert count(["apple"]) == {"apple": 1}


def test_count_3() -> None:
    """Mock unit tests for dictionaries."""
    assert count(["apple"]) == {"apple": 1}


def test_alphabetizer_1() -> None:
    """Mock unit tests for dictionaries."""
    assert alphabetizer(["apple"]) == {"a": ["apple"]}


def test_alphabetizer_2() -> None:
    """Mock unit tests for dictionaries."""
    assert alphabetizer(["apple"]) == {"a": ["apple"]}


def test_alphabetizer_3() -> None:
    """Mock unit tests for dictionaries."""
    assert alphabetizer(["apple"]) == {"a": ["apple"]}


def test_update_attendance_1() -> None:
    """Mock unit tests for dictionaries."""
    adict: dict[str, list[str]] = {}
    update_attendance(adict, "Monday", "Vrinda") 
    assert adict == {"Monday": ["Vrinda"]}


def test_update_attendance_2() -> None:
    """Mock unit tests for dictionaries."""
    adict: dict[str, list[str]] = {}
    update_attendance(adict, "Monday", "Vrinda") 
    assert adict == {"Monday": ["Vrinda"]}


def test_update_attendance_3() -> None:
    """Mock unit tests for dictionaries."""
    adict: dict[str, list[str]] = {}
    update_attendance(adict, "Monday", "Vrinda") 
    assert adict == {"Monday": ["Vrinda"]}
